Game code efficiancy update, base2d doc update, other small fixes
Game code updated in tiles to not getMoves and getJumps every frame, only updates if (from tiles) either self.moves or self.jumps is None, also added a function to be called by the game board specifically to reset move and jump information. Updated game board to add a function to be called by tiles to indicate a move has been made, and the game board then tells all tiles to re-evaluate move information as described above. Added a bit of documentation to the base2d module, as the State class previously had no documentation about what it does. Also updated the game to not have the world be a global variable, as to not allow AIs to manually call functions to be executed without permission (pretty much eliminates AIs from modifying game board tile data without properly completing a move)