Man, a lot of changes. Read extened discription.
In the base game, there are now modifications in place that allow AIs to define what the starting player should be, define player names, and make the game quit instead of starting another game. Also, AIs can pass their turn (pretty much just re-sends board data update and calls the AI to make a turn again), and in extreme cases cause the game to close. AIs have two additional functions that can be called now, one that is called when the game quits, and one to inform the AI about whether the move it just tried to make was successful or not.

In the ExampleCheckersAI, the new functions described above were implemented, even though they just pass.

Added HackyCheckersServer.py and HackyCheckersClientNotAnAI.py
The server is a fairly simple threaded (mostly) socket server, that when run, binds the the host computer's outwards network IP address on port 8673, which from my research is a unused port for anything in the public domain, which is connected to by exactly two clients. No more, no less.
The client's filename is important, as it triggers the game to THINK it's an AI and load it. In reality, the so called "AI" is really a client for a chess server, described above, allowing two players to play a game of checkers over the network, instead of doing it all one one machine.

If you have any questions, please ask me. Thank you for your time.